fdc/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
dataclasses-json==0.5.7
Django==4.2
django-postgres-extra==2.0.8
djangorestframework==3.14.0
drf-yasg==1.21.5
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
langchain==0.0.162
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openai==0.27.6
openapi-schema-pydantic==1.2.4
packaging==23.1
prompt-toolkit==3.0.38
protobuf==3.20.3
psycopg2==2.9.6
psycopg2-binary==2.9.6
pydantic==1.10.7
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
redis==4.5.5
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
SQLAlchemy==2.0.13
sqlparse==0.4.4
tenacity==8.2.2
tqdm==4.65.0
typing-extensions==4.5.0
typing-inspect==0.8.0
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
yarl==1.9.2
